+++ exportplan/helpers.py
@@ -21,7 +21,7 @@ def get_exportplan_rules_regulations(sso_session_id):
 def get_madb_country_list():
     airtable = Airtable('appcxR2dZGyugfvyd', 'CountryDBforGIN')
     airtable_data = airtable.get_all(view='Grid view')
-    country_list = [c['Country'] for c in [f['fields'] for f in airtable_data]]
+    country_list = [c['country'] for c in [f['fields'] for f in airtable_data]]
     return list(zip(country_list, country_list))
 
 
@@ -29,8 +29,8 @@ def get_madb_commodity_list():
     airtable = Airtable('appcxR2dZGyugfvyd', 'CountryDBforGIN')
     commodity_name_set = set()
     for row in airtable.get_all(view='Grid view'):
-        commodity_code = row['fields']['Commodity code']
-        commodity_name = row['fields']['Commodity Name']
+        commodity_code = row['fields']['commodity_code']
+        commodity_name = row['fields']['commodity_name']
         commodity_name_code = f'{commodity_name} - {commodity_code}'
         commodity_name_set.add((commodity_code, commodity_name_code))
     return commodity_name_set