+++ exportplan/urls.py
@@ -7,5 +7,6 @@ app_name = 'exportplan'
 urlpatterns = [
     path('start/', views.ExportPlanStartView.as_view(), name='start'),
     path('create/', views.ExportPlanCreateView.as_view(), name='create'),
+    path('section/target-markets/', views.ExportPlanTargetMargetsView.as_view(), name='target-markets'),
     path('section/<slug:slug>/', views.ExportPlanSectionView.as_view(), name='section'),
 ]